Reorganise pages imports
