+++ cypress/e2e/right-slot.cy.ts
@@ -1,6 +1,5 @@
 import { breakpoints } from '@guardian/source-foundations';
-import { articles } from '../fixtures/pages/articles';
-import { liveblogs } from '../fixtures/pages/liveblogs';
+import { articles, liveblogs } from '../fixtures/pages';
 
 describe('right slot on pages', () => {
 	[...articles, ...liveblogs].forEach(({ path, adTest }) => {
